This is a weather analysis project developed in Python
This project can be used in analysing the current weather condition. I have used the open weather API to get the cuttent data. I have wrote the code in a very flexible manner so anyone can reuse this by putting their API details and API Key. The code is a ipython note book which contains the comments and explaination of each steps and logic I have used.